Deep copy in load_qtl
Reaised we were only laoding in the final band's QTL, and that the orbitals dictionary was absolutely scuffed. Turns out, was doing a refernce copy of a single dictionary, rather than a deep copy to create new dictionaries every time I made a new band in the QTL dictionary. This was bad. Have fixed it with copy module's copy.deepcopy() function.